Update post.py
Updated to utilize the new pytumblr system. (this works)

After 3 pushes it returns 'bad request'. This is something I am trying to figure out, I believe this is a maximum amount able to be submitted via api within a timeframe but it does not appear to be listed anywhere. I will continue to work on the API calls to identify where this issue comes from. Hopefully there will be a result for this soon.