Update post.py
Added new random caption generator. Also made the caption use one of 3 available systems (for the time being). This is so that we can attempt to rewrite the systems into one system later on. The new random generation system is pretty, and is purely random. All phrases generated are creative in nature and do not limit the system.

Removed 'approved_types' tuple. No longer required in the current state of the script.

Added broken_files array. When a file returns an error of 8009, it will get put into this array.
No further attempts at this file will be made until the push_array is clear.

Added exceeded_files array. When a file returns an error of 8011, it will get put into this array.
No further attempts at this file will be made until the date changes. Because 8011 is the error for exceeded allowed posts (video only)